requirements.txt
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.28.0
django-bootstrap3==7.1.0
django-categories==1.4.3
django-configuration==0.0.3
django-configurations==2.0
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-mptt==0.8.6
django-pagedown==0.1.1
django-polymorphic==1.0.2
django-taggit==0.21.3
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.7
markdown2==2.3.1
oauthlib==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
python3-openid==3.0.10
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
unicode-slugify==0.1.1
wcwidth==0.1.7
